world-drive: refactor backend structure & coordinate type
