+++ backend/world-drive/src/types/Coordinate.ts
@@ -1,0 +1,7 @@
+/**
+ * GPS Coordinate interface representing a specific point on a two-dimensional map by latitude and longitude.
+ */
+export interface Coordinate {
+    latitude: number
+    longitude: number
+}
