+++ backend/world-drive/src/types/WeatherData.ts
@@ -1,0 +1,8 @@
+/**
+ * WeatherData interface representing the weather information for a specific city.
+ */
+export interface WeatherData {
+    degreesCelsius: number
+    weatherDescription: string
+    cityName: string
+}
