--- backend/world-drive/types/sharedTypes.ts
@@ -1,14 +1,0 @@
-/**
- * GPS Coordinate type representing a specific point on a two-dimensional map by latitude and longitude.
- */
-export interface Coordinate {
-    lat: number
-    lng: number
-}
-
-export interface WeatherData {
-    degreesCelsius: number
-    weatherDescription: string
-    cityName: string
-}
-
