+++ frontend/pse-cars/app/world-drive/components/WorldDriveComponent.tsx
@@ -5,7 +5,7 @@ import WeatherComponent from "@/app/world-drive/components/WeatherComponent";
 import MapboxComponent from "@/app/world-drive/components/MapboxComponent";
 
 export default function WorldDriveComponent() {
-    const [position, setPosition] = useState<Coordinate>({lat: 48.8603192, lng: 9.1780495})
+    const [position, setPosition] = useState<Coordinate>({latitude: 48.8603192, longitude: 9.1780495})
 
     useEffect(() => {
         const handleCoordinate = (coordinate: Coordinate) => setPosition(coordinate)
