+++ frontend/pse-cars/app/world-drive/service/currentPositionApiService.ts
@@ -1,6 +1,6 @@
 export type Coordinate = {
-    lat: number
-    lng: number
+    latitude: number
+    longitude: number
 }
 
 type CoordinateCallback = (coordinate: Coordinate) => void
